+++ src/graphql/resolvers.ts
@@ -1,12 +1,7 @@
 import {GraphQLError} from "graphql";
 import * as status from "http-status";
-import {Sequelize} from "sequelize";
 import dataaccess from "../lib/dataaccess";
-import {Chatroom} from "../lib/dataaccess/Chatroom";
-import * as models from "../lib/dataaccess/datamodels";
-import {Post} from "../lib/dataaccess/Post";
-import {Profile} from "../lib/dataaccess/Profile";
-import {User} from "../lib/dataaccess/User";
+import * as models from "../lib/dataaccess/models";
 import {NotLoggedInGqlError} from "../lib/errors/graphqlErrors";
 import globals from "../lib/globals";
 import {InternalEvents} from "../lib/InternalEvents";
@@ -21,8 +16,7 @@ export function resolver(req: any, res: any): any {
     return {
         async getSelf() {
             if (req.session.userId) {
-                const user = await models.SqUser.findByPk(req.session.userId);
-                return user.profile;
+                return models.User.findByPk(req.session.userId);
             } else {
                 res.status(status.UNAUTHORIZED);
                 return new NotLoggedInGqlError();
@@ -32,8 +26,7 @@ export function resolver(req: any, res: any): any {
             if (handle) {
                 return await dataaccess.getUserByHandle(handle);
             } else if (userId) {
-                const user = await models.SqUser.findByPk(userId);
-                return user.user;
+                return models.User.findByPk(userId);
             } else {
                 res.status(status.BAD_REQUEST);
                 return new GraphQLError("No userId or handle provided.");
@@ -49,8 +42,7 @@ export function resolver(req: any, res: any): any {
         },
         async getChat({chatId}: { chatId: number }) {
             if (chatId) {
-                const chat = await models.SqChat.findByPk(chatId);
-                return new Chatroom(chat);
+                return models.ChatRoom.findByPk(chatId);
             } else {
                 res.status(status.BAD_REQUEST);
                 return new GraphQLError("No chatId given.");
@@ -110,8 +102,8 @@ export function resolver(req: any, res: any): any {
         async vote({postId, type}: { postId: number, type: dataaccess.VoteType }) {
             if (postId && type) {
                 if (req.session.userId) {
-                    const post = await models.SqPost.findByPk(postId);
-                    return await post.post.vote(req.session.userId, type);
+                    const post = await models.Post.findByPk(postId);
+                    return await post.vote(req.session.userId, type);
                 } else {
                     res.status(status.UNAUTHORIZED);
                     return new NotLoggedInGqlError();
@@ -138,8 +130,8 @@ export function resolver(req: any, res: any): any {
         },
         async deletePost({postId}: { postId: number }) {
             if (postId) {
-                const post = (await models.SqPost.findByPk(postId)).post;
-                if ((await post.author()).id === req.session.userId) {
+                const post = await models.Post.findByPk(postId, {include: [models.User]});
+                if (post.rAuthor.id === req.session.userId) {
                     return await dataaccess.deletePost(post.id);
                 } else {
                     res.status(status.FORBIDDEN);
@@ -200,8 +192,8 @@ export function resolver(req: any, res: any): any {
                 return new NotLoggedInGqlError();
             }
             if (sender && type) {
-                const profile = new Profile(req.session.userId);
-                await profile.denyRequest(sender, type);
+                const user = await models.User.findByPk(req.session.userId);
+                await user.denyRequest(sender, type);
                 return true;
             } else {
                 res.status(status.BAD_REQUEST);
@@ -215,8 +207,8 @@ export function resolver(req: any, res: any): any {
             }
             if (sender && type) {
                 try {
-                    const profile = new Profile(req.session.userId);
-                    await profile.acceptRequest(sender, type);
+                    const user = await models.User.findByPk(req.session.userId);
+                    await user.acceptRequest(sender, type);
                     return true;
                 } catch (err) {
                     globals.logger.warn(err.message);