+++ src/lib/dataaccess/models/ChatMember.ts
@@ -1,0 +1,14 @@
+import {Column, ForeignKey, Model, Table,} from "sequelize-typescript";
+import {ChatRoom} from "./ChatRoom";
+import {User} from "./User";
+
+@Table({underscored: true})
+export class ChatMember extends Model<ChatMember> {
+    @ForeignKey(() => User)
+    @Column
+    public userId: number;
+
+    @ForeignKey(() => ChatRoom)
+    @Column
+    public chatId: number;
+}