+++ src/lib/dataaccess/wrappers.ts
@@ -1,5 +1,0 @@
-export {User} from "./User";
-export {Chatroom} from "./Chatroom";
-export {Post} from "./Post";
-export {Profile} from "./Profile";
-export {ChatMessage} from "./ChatMessage";